minor fixes to code and improved tests
